Add RandomOperand flags: doNotCheckAccuracy and doNotConnect.
This CL provides the following fixes:

* Should not check accuracy of TOPK_V2 second output

We should not check the second output as the sorting is not required
to be stable.

The second output of TOPK_V2 is marked as doNotCheckAccuracy and
doNotConnect.

* Handle NaN/Infinity float values in RGG

Some operations that may produce invalid floating point values,
e.g. NaN, Inf. We should not connect the output tensor of such
operation to the input of another operation, as some floating point
operations are undefined upon NaN and Inf.

A complete list of such operations:
* ANEURALNETWORKS_DIV
* ANEURALNETWORKS_LOG
* ANEURALNETWORKS_POW
* ANEURALNETWORKS_SQRT
* ANEURALNETWORKS_RSQRT
* ANEURALNETWORKS_L2_NORMALIZATION
* ANEURALNETWORKS_REDUCE_PROD

The output of these operations are marked as doNotConnect.

Bug: 134800514
Bug: 134753636
Test: NNT_static_fuzzing
Change-Id: Id71244c5fe3b26ba9a483c361867a3021c7de20e