Fix bidirectional sequence RNN crash when auxiliary input is omitted
The crash would happen if auxiliary inputs were provided with a shape
with less than 3 dimensions. This is the case when marking optional
input as omitted.

The tests weren't catching this behaviour because they set three- or
two-dimensional shape consisting of zeros instead of just
one-dimensional zero. The change modifies the tests so that they use
one-dimensional shape.

Fix: 135242001
Test: NeuralNetworksTest_static
Change-Id: I47168d6a59445d2591a3c25ce7bbd4e17e6b2547