Ensures correct LOG_TAG is used by NN_RET_CHECK_FAIL macro
Current implementation of the NN_RET_CHECK_FAIL macro relies on
the FalseyErrorStream class declared in Utils.h. The class appears
in every including file with an implementation depending
on the value of the LOG_TAG captured by the including source.

For example in libneuralnetworks_common there are 55 definition of the
class

scanelf -g -s ".*FalseyErrorStream.*" \
 -A libneuralnetworks_common.a | grep FalseyErrorStream | c++filt

This change places every FalseyErrorStream into an anonymous namespace
so that every file will be using its own version with the correct
LOG_TAG.

Bug: 131095310
Test: NeuralNetworksTest_logtag
Test: NeuralNetworksTest_operations
Test: NeuralNetworksTest_utils
Change-Id: Iceb6de426e8145338c4208326357df4db12153e0