Enable IWYU for cpp files and formatting for bp files in NNAPI presubmit hook
This enables a presubmit hook to check if the proper header files are
included. It looks like this lint currently only checks for which STL
types and functions are missing, and does not seem to check for (1)
if a header is unnecessarily included, or (2) if any non-STL headers
are missing. We can explore what other lints we want to enable.

This CL also enables formatting on bp files.

Bug: N/A
Test: none
Change-Id: I20818e5304f46bcb174e4480333e4e29a9136133