Reorganize model slicing data structures: Replace IModelSlicer with MetaModel
The previous data structure organization employed an interface class
IModelSlicer that was passed around along with a Model; a
concretization of that class would manufacture and cache slices.

The new data structure organization employs a concrete class
MetaModel that that is passed around in place of a Model;
this class exposes the original Model, and manufactures and caches slices.

Previously, ANeuralNetworksModel_getSupportedOperationsForDevices() did
not perform slicing; now it does.

Also replace uses of std::ostrstream with std::ostringstream.

Bug: 132378048

Test: NeuralNetworksTest_static (with debug.nn.strict-slicing 1)
Test: NeuralNetworksTest_static_fuzzing --gtest_filter=TestRandomGraph/RandomGraphTest.* (with debug.nn.strict-slicing 1)

Change-Id: Iff694b9c8391f601592729a248bff2adeba2332f