Cleanup Manager's Device class
This CL includes the following changes:
* Change instances of "const char*" to "const std::string&"
* Make DriverDevice members constant
* Change listManifestByInterface to getAllHalInstanceNames

Elaborating on the last point, previously the NNAPI runtime would get
a handle to the defaultServiceManager1_2, then call its method
listManifestByInterface in order to get the names of all HAL instances.
getAllHalInstanceNames is a new, simpler HIDL call that performs both
of these actions and directly returns the names.

Fixes: 76116804
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: Icf98281b033fb7bcd8b1d1bb4d02d398b0979081