Combine synchronous and asynchronous HAL executions
Prior to this CL, VersionedIPreparedModel contained both execute and
executeSynchronously methods. However, which method was called was
determined by a system property. This CL consolidates them to a single
call "execute", which includes a "preferSynchronous" flag to select the
synchronous/asynchronous behavior.

To simplify this refactor, the asynchronous "execute" call was changed
to be synchronous. This contains and manages the protections for the
asynchronous HAL call within executeAsynchronously method, and makes the
behavior of VersionIPreparedModel easier to reason about.

This CL also removes VersionedIDevice's and VersionedIPreparedModel's
equality and inequality operators because they are not used, adds
missing includes, and adds extra documentation.

Bug: 122316159
Fixes: 132322612
Fixes: 132322455
Fixes: 132322234
Fixes: 132323703
Test: mma
Test: atest NeuralNetworksTest_static
Test: atest CtsNNAPITestCases
Change-Id: If03a83d497d8e35e82aa963e5aeefa7e30c96993