Replace libbinder dependency with libneuralnetworks_packageinfo
libbinder communication from APEX is not allowed, because it doesn't
form a stable way to communicate. After an update, if system libbinder
and APEX libbinder are out of sync, then they may fail to work with each
other.

libbiner usage is moved into libneuralnetworks_packageinfo library,
which is a stable API library.

Test: flashed crosshatch / run NNAPI tests
Bug: 139282353
Change-Id: Ibe7e01b7f3014aed8390091ff7250611a32bb2ee