Simplify IDevice reboot logic
HIDL allows a service to be retrieved with two functions:
* <Interface>::getService — blocks until service is retrieved
* <Interface>::tryGetService — immediately returns service or nullptr

Currently, the NN runtime retrieves the service in three different
places:
1) When the runtime first starts, <Interface>::getService is used to
   acquire all services
2) When the object is dead, <Interface>::tryGetService is used to
   attempt to reacquire the service, but will quickly resume if the
   service is still rebooting
3) When the client calls ANNDevice_wait, <Interface>::getService is used
   to block until the service is active again

This CL simplifies the IDevice reboot logic by changing these static
class functions to dependency injection. Specifically, VersionedIDevice
now retrieves a handle to the IDevice object through a DeviceFactory
function that is passed in when the VersionedIDevice object is created.
This function can either operate as blocking or nonblocking to support
all use-cases described above, and makes it easier to test the
VersionedIDevice recovery code.

Bug: 139189546
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: I012eb4df6f09f98bdfbd0835457ba98bc22d906e