Create VersionedIDevice with constant members
This CL relocates (but does not change, with the exception of
removing a log message from "getTypeFunction") the initialization
functions higher in the file to make them come before device
initialization code.

This CL also changes the device initialization code to call these
initialization functions before the VersionedIDevice object is
created, in order to make these cached initialization values
constant. VersionedIDevice's constant members' names are changed to
be prefixed with a "k".

The initialization functions include:
* getCapabilitiesFunction
* getSupportedExtensionsFunction
* getTypeFunction
* getVersionStringFunction
* getNumberOfCacheFilesNeededFunction

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: Id93fc0650a9163313cdb62fc38d937060ebc4805