Split VersionedIDevice initialization function by version
This CL splits the HIDL IDevice methods involved in initialization
into different versioned free function calls. This CL additionally
simplifies these initialization method calls by removing the
recoverable logic from these calls. Prior to this CL, these
initialization calls would only make use of the recoverable path if the
driver managed to crash and reboot itself between the time the IDevice
handle was originally retrieved and the call is made to the various
initialization methods. This window of time is very small, so there did
not seem to be much benefit from making these initialization methods
recoverable. Instead, they are removed in this CL for simplicity.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: I391f7b3ecf4c44b204140d0e943f4179fa5aec60