Make DriverDevice::initialize return false for any error
Prior to this CL, DriverDevice::initialize attempted to gracefully
recover from the case when the NN HAL driver fails to return
getNumberOfNeededCacheFiles or when the needed number of cache files
exceeds the specified max number. It recovers gracefully by defaulting
the needed number of cache files to 0, indicating that caching is not
supported. This CL changes the behavior such that this is considered
an initialization failure, causing the calling code not to use that
DriverDevice object.

This CL also changes the TestCompilationCaching to match this new
initialize behavior.

Fixes: 132322236
Test: mma
Test: NeuralNetworksTest_static
Test: CtsNNAPITestCases
Change-Id: I97e40be2553e06f4c97727184286000af2529d45
Merged-In: I97e40be2553e06f4c97727184286000af2529d45
(cherry picked from commit 1807e65239ba129d071ba1528466c98021a00460)