Switch to libcrypto_static for com.android.neuralnetworks APEX
This reduces com.android.neuralnetworks APEX size and memory use.
After moving libneuralnetworks.so to an APEX module, it can't use
the system libcrypto.so. This results in a copy of the libcrypto.so
being added to the com.android.neuralnetworks APEX. By switching to
static version of libcrypto, we link in just the required SHA functions.

Test: Flashed pixel device & run NNAPI tests.
Bug: 140440028
Change-Id: I605145581287ab4b2b3f23bbab5c69f6e8e13762