Make sample drivers not able to prepare models they don't support
Before this change sample drivers that only support a subset of our ops
(FloatFast, FloatSlow, Quant and Minimal) only overrode
getSupportedOperations_1_3 method of the full SampleDriver but not
prepareModel_1_3. Our VTS code for generated tests tries to prepare a
model even if a driver says that it doesn't support some/all of the ops in the
model. Since the drivers used full implementation of prepareModel_1_3,
they succeeded at preparation and computation continued until
successful values comparison. So, no VTS tests were skipped for these
sample drivers.

This change overrides prepareModel_1_3 for the partial sample drivers so
that they are not able to prepare a model they said they can't prepare.
After this change, some of the VTS tests for the partial sample drivers
will be skipped.

Bug: 137828994
Test: VtsHalNeuralnetworksV1_3TargetTest
Change-Id: I3b5df05e2fe98dd1dc2113f58c7409171d94e505