Introduce libneuralnetworks_static
The library is a static variant of the existing libneuralnetworks. It is
introduced because the existing one isn't available to the platform due
to the apex_available property. (the property was set in order to
prevent installation of the lib to /system/lib)

The problem was that a few number of nn tests are statically linking the
library. Since the nn tests are outside of the APEX, they need the
platform variant of the library. In order to solve the problem, the
static variant is introduced just for the tests. Since it's a static
library, it still isn't installable to /system/lib.

Bug: 147158681
Test: mma under frameworks/ml/nn
Change-Id: I61a56bf3bb9fde87d0268d911f4a22ff0febafd9