Relax setTimeout duration parameter
Prior to this CL, the NNAPI QoS deadline is a strict deadline: a task
with a deadline must either complete or abort by the deadline. To
avoid as much overhead as possible, it is requested for Android R to
have a relaxed deadline, where the task *may* be aborted after the
deadline has passed or when the task is estimated to take too long.

Because the deadline is now relaxed, the querying functions
"supportsCompilationTimeout" and "supportsExecutionTimeout" can be
removed.

Bug: 149766387
Test: mma
Test: CtsNNAPITestCases
Test: NeuralNetworksTest_static
Change-Id: Ib76bdd8697686b8d6f3414d3742998c67b2f1f64