Change PreparedModel execute interface to use vector instead of MemoryTracker.
MemoryTracker is useful in ExecutionBuilder when creating the execution
requests by setInput/OutputFromMemory. However, when computing the
graph, the memory objects should already been frozen.

This CL changes the interface of PreparedModel::execute to use vector of
memories instead of MemoryTracker. The purpose of this change is to
prepare for ag/10827131, where some of the memories need to be replaced
with BLOB mode AHardwareBuffer.

Bug: 152208838
Test: NNT_static
Change-Id: I4f13134fa74a4bc6f214fad358131da605cc9850