Compute RGG golden results from FP model.
Before this CL, the golden results in RGG tests are computed from
nnapi-reference with the actual data type. This CL changes the logic to
compute the golden results from an equivalent float32 model to avoid
bias if possible.

Currently only the golden results of single-operation models are
computed from fp32 models. This is because there is no multi-operation
quant model in RGG tests yet.

Fixes: 150805665
Test: NNT_static
Change-Id: I4e268d3aadeebdcdae1d7010a4d0564f45b11b82