Add rank check to concatenation op
Without this check concatenation would crash if passed 5D tensors as
inputs.

Also:
* Add a folder for CTS only tests and modify Android.bp accordingly.
* Add concatenation test with invalid rank to CTS.
* Modify createModel and GeneratedTests::execute functions to handle
  validation failures for models with exepectFailure flag.

The test is added only to CTS because we will very likely relax this
rank requirement in the future. If this test was added to VTS, future
drivers would have to contain logic to reject models with larger ranks
from clients using older versions. In order to make driver development
easier, we only introduce this test to CTS. Once we relax the rank
requirements, we will remove corresponding tests.

Fix: 139957680
Test: NNTest_static and VtsHalNeuralnetworksV1_3Target
Change-Id: I6d3b825360e055934a4e2d44e50b8fdc83b958e3