Fix the issue that INSTANCE_NORMALIZATION is not numerically stable.
Prior to this CL, the variance is computed as:

  sigma^2 = sum(x^2) / len - mean^2

This is not numerically stable because sum(x^2) may go very large and
lose precision.

This CL computes the variance by

  sigma^2 = sum((x - mean)^2) / len

This is more stable especially when sigma^2 is much smaller than mean^2.

Additionally, this CL disallows invalid FP values for generated tests.

Fixes: 151360275
Test: NNT_static
Change-Id: I2842dd44553ef9ca4876f198763d741640e5178c