Fix control flow partial CPU fallback bug
Before this change, ExecutionPlan::Controller::mLastStepIndex is only
updated at the beginning of the ExecutionPlan::next() call. When
interpreted control flow is involved, this call might result in several
logical steps being processed until an ExecutionStep is reached.

When an interpreted CF loop terminates successfully but the next step is
an ExecutionStep that fails, ExecutionPlan::fallback() restarts from
mLastStepIndex, which happens to be the index of the WhileStep. This
results in the loop being executed again. This happens forever.

This change fixes the bug by updating mLastStepIndex when
nextCompound(const ExecutionStep* ...) is called.

This change renames mLastStepIndex to mFallbackNextStepIndex to reflect
its usage.

Fix: 152623151
Test: NNT_static
Change-Id: Iff3b68f541adb8facf078e968abac2a6ae8e74c1