libneuralnetworks_headers is marked as available to platform
The header lib is depended on by libneuralnetworks_common which is
available to platform. With b/153073816, we will be checking the
availability to the platform as well, i.e. platform-available module
dedpending on platform-unavailable module will cause a build error.

Prepare for the change by correctly marking that the header lib is also
available to the platform.

Bug: 153073816
Test: m

Change-Id: I31a198068c98e7d6b01b7e2cd6f43e54769f28b1