Fix NNAPI TimingTests
This CL fixes the ordering of sequences in the test. Specifically:
1) Causes the spin-loop in waitForExecutionToBegin to spin while the
   execution has not yet started
2) Adds dummyExecution() to the driver implementations of FAIL_LAUNCH
3) Skips getDurationWhileRunning for all DriverKind::CPU test cases

Further, this CL adds dummyExecution() to
TestPreparedModelLatest::executeFenced to ensure the call to
executeFenced will not complete until after the test has called
getDuration. This brings executeFenced in line with the other execution
methods which call dummyExecution().

Finally, this CL has the following two minor changes:
1) code that was previously running from a detached thread in
   executeFenced now instead runs synchronously
2) a for-loop that iterated by constant reference now instead iterates
   by constant value

Fixes: 155215099
Test: mma
Test: NeuralNetworksTest_static --gtest_filter=Fenced/TimingTest*
Change-Id: I69a7448367ce8353791c9462d144122b6c913802