Fix caching bug in quantization coupling test
Execute function in our tests among other things makes sure that the
models work with caching enabled. Caching token and directory are
constant and therefore the same for all tests. This is usually not a
problem since the tests check only one model with a set of different
configs but this was not the case for QuantizationCouplingTest suite.
This suite ran signed and unsigned models in one test making them share
the same caching token and directory which is an incorrect use of
caching API.

This CL solves this problem by removing the execution of an unsigned
model from the test. This is safe since the unsigned model is still
verified in GeneratedTests suite.

Bug: 155604227
Test: NNTest_static
Change-Id: I6aa2c8cd3fa41e37a5de53259d1fcc784f327093
Merged-In: I6aa2c8cd3fa41e37a5de53259d1fcc784f327093
(cherry picked from commit b66fed41537e4f7cff0f4937aba30dff57cee41f)