UPSTREAM: Fix caching bug in quantization coupling test
Execute function in our tests among other things makes sure that the
models work with caching enabled. Caching token and directory are
constant and therefore the same for all tests. This is usually not a
problem since the tests check only one model with a set of different
configs but this was not the case for QuantizationCouplingTest suite.
This suite ran signed and unsigned models in one test making them share
the same caching token and directory which is an incorrect use of
caching API.

This CL solves this problem by removing the execution of an unsigned
model from the test. This is safe since the unsigned model is still
verified in GeneratedTests suite.

Upstream Bug: 155604227
Test: NNTest_static
(cherry picked from commit 65432baae00a8bea7a01627cd533daf478acdb44)

BUG=b:157855577
TEST=FEATURES="test" emerge-octopus chromeos-base/aosp-frameworks-ml-nn

Change-Id: I303704106a9843f3057a52106185a3f1f8915623
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2225001
Reviewed-by: Alan Green <avg@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>