Fix null ptr dereference in cpu implementation of operations
Operations fixed:
* MEAN
* ARGMIN/ARGMAX
* STRIDED_SLICE

The operations would crash when provided with inputs that resulted in an
empty output shape.

The change fixes the bug by making the operations output a tensor of
size [1] in this case.

Also, update the documentation to clarify this behaviour and move
squeeze operation test to appropriate spec directory.

Bug: 155508675
Bug: 155660285
Bug: 155508675
Bug: 155238914
Test: NNTest_static
Change-Id: Ia865c26021dd4d781659957049dd567beeaeae99