Avoid sending ahwb requests to 1.0 and 1.1 drivers.
AHWB buffer support is a feature introduced in 1.2. Prior to this
CL, requests with AHWB as memory pool will be sent to 1.0 and 1.1
drivers. This CL modifies the compliance check and avoids sending
ahwb to 1.0 and 1.1 drivers. Using AHWBs on compilations with 1.0
and 1.1 drivers will result in a CPU fallback.

Bug: 155686276
Test: NNTS
Change-Id: Ib00a2d0d24d4a8c385b5992a1168e20c4a6bb786