Use scales representable in fp16 for fp16 casting ops in RGG.
The scale in NNAPI is always a fp32 data type. As there are accelerators
that only support fp16 data type, there is precision loss when
executing a casting operation that converts a fp16 buffer to/from a
quantized buffer. In the worst case, the precision loss in scale could
be up to 0.5 ULP, and results in 0.5 * 255 = 127.5 ULP difference in the
final result. This is much larger than the tolerance of these operations
in RGG. This CL sets the scale of such operations to be always representable
in fp16 to avoid potential precission loss.

Additionally, relax the MSE requirement on DEQUANTIZE.

Bug: 155842363
Test: NNT_static_fuzzing
Change-Id: I511ddb6753be1dabac36e1423820a03a4f28641e