ml/nn: Add common_operations lib
Adds an initial version of the common_operations library. Most of the
code compiles, but tests are not yet running.

More investigation is needed into why some files do not compile, but it
seems that at least some of the problem is due to differences in the
TfLite API.

BUG=b:157391897
TEST=It compiles!

Cq-Depend: chromium:2219752
Change-Id: I10704343a75752a7535d61a6efd6d5b7c142cb5a
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2214726
Tested-by: Alan Green <avg@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Alan Green <avg@chromium.org>