Relaxed LSH_PROJECTION: do not convert the hash tensor as internal
LSH_PROJECTION is very sensitive to the value of the hash tensor. Prior
to this CL, AllInputsAsInternalCoverter will convert the hash tensor to
internal by introducing a dummy ADD operation. Under relaxed execution
mode, the small precision loss in ADD will result in a significant
difference in the final result. This CL prevents the hash tensor from
being converted to internal in relaxed precision tests.

Additionally, this CL removes a redundant variation in
lsh_projection_float16.

Fixes: 155962587
Test: NNT_static
Change-Id: Id5522b4949a4e3ab4801537e8eb747a25f0cd0e8