Fix fuzzer buffer sizes
Prior to this CL, the NN fuzzers would randomly generate a buffer for
each operand. However, NNAPI validation immediately failed if the
provided buffer did not match the expected buffer's size. This made it
difficult for the fuzzer to converge on "interesting" test cases. This
CL forces each buffer to match its operand's shape (assuming it's a
reasonable size). Although this does "skip" some validation code in most
cases, this enables the fuzzer to more quickly converge on more
interesting test cases.

Bug: N/A
Test: mma
Test: libneuralnetworks_fuzzer
Test: libneuralnetworks_driver_fuzzer
Change-Id: I46ca9a5c1718f25c8c1b071196ecc13800742074