Simplify NN fuzz tests
This CL combines the common aspects of libneuralnetworks_fuzzer and
libneuralnetworks_driver_fuzzer into the library
libneuralnetworks_fuzzer_harness. To implement a new fuzz test, a
developer only needs to implement the function:
  void nnapiFuzzTest(const TestModel& testModel)

This CL also introduces a soong default
libneuralnetworks_fuzzer_defaults so that a target can more easily
import the necessary libraries and the NNAPI corpus.

Bug: N/A
Test: mma
Test: libneuralnetworks_fuzzer
Test: libneuralnetworks_driver_fuzzer
Change-Id: Ic9bdca6369f6dd809816a5a24223a6ed2f55f6ca