nn: Enable building of runtime/Manager.cpp
Added the define NNAPI_CHROMEOS to selectively build on Chrome OS, to
keep the file relatively intact and make it easier to merge in changes
from upstream.

ifdef'd out the HAL device enumeration, which we will need to replace
with our own mechanism for loading the HAL from a shared library, as
nest did.

BUG=b:158791375, b:157518468
TEST=emege-octopus chromeos-base/aosp-frameworks-ml-nn

Change-Id: Id2d897c44e091ba3b32f810efbaf2783abba1692
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2245842
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>