nn: register SampleDriverFull into DeviceManager
SampleDriverFull is an implementation of IDevice that should allow
validation of nnapi. At the moment it is hardcoded as the only device
registered in the system. Next steps would be to replace this with a
more dynamic mechanism that will allow easily changing which driver is
used.

BUG=b:158791375
TEST=FEATURES=test emerge-kukui chromeos-base/aosp-frameworks-ml-nn

Cq-Depend: chromium:2248042
Change-Id: I28b9daf7a2318f84f0e696db62c08cd5f2b492d4
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2246507
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>