Do not schedule unknown size control flow operations to real devices
This feature would likely have very limited vendor support in Android R.
It's too late to add IF or WHILE tests to the 1.3 VTS where an inner or
outer input or output operand of a control flow operation has a type
that is not fully specified. To avoid exercising untested behaviour, we
have decided to disallow this at the HAL level. This change ensures that
we do not schedule such operations for execution on any device other
than the CPU device.

See http://b/159076604#comment5 and http://b/132458982#comment63.

Bug: 159076604
Bug: 132458982
Test: NNT_static
Change-Id: Ic5f864e6129acc4208b7751b5b182ae30a39f0a4