nn: Ensure that symbols have the correct visibility.
By default symbols have visibility internal, which means that there was
nothing exported from the shared library. We need to mark the runtime
symbols as visible, as they are loaded by tflite. This also ensure that
everything is linked into the .so correctly.

- Add missing link dependencies.
- We needed to add sample_driver, but I don't think we should. Follow up
work is to remove this.

BUG=b:155344260
TEST=emerge-betty chromeos-base/aosp-frameworks-ml-nn
TEST=cros deploy to betty vm
TEST=ml_cmdline --nnapi passes.

Change-Id: Icd77297c36786f699a17e4d4c81dc78c531be1b4
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/frameworks/ml/+/2256101
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>